CMake: Split static library exports into their own export file
Summary:
This is to better support distros which split the static libraries into
their own package.

The current problem is that any project the includes LLVMConfig.cmake
will fail to configure unless the static libraries are installed.  This
is because LLVMConfig.cmake includes LLVMExports.cmake, which throws an
error if it can't find files linked to one of the exported targets.

This patch resolves the problem by putting the static library targets
into their own export file, LLVMStaticExports.cmake.  This file
is optionally included by LLVMConfig.cmake, so distros can put this
new file in their static library package to make  LLVMConfig.cmake
no longer depend on these libraries when they are not installed.

Reviewers: beanz, mgorny, chapuni

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D32668