Update git-llvm script to push to GitHub
Summary:
Note: This patch should not be pushed until SVN has become read-only.
It should be the first patch committed directly to GitHub.

This patch updates git-llvm to check for merge commits and then push
changes to GitHub if none are found.  All logic related to SVN has been
removed.

I need help testing this to make sure it is robust and that the merge
commit checks can cover every case.  To test only the merge checks,
just run git-llvm push --dry-run.

If you want to also test pushing chagnes, change the GIT_ORG and
GIT_REPO variables to one of your personal github repositories.
You will also need to enable the branch protections and the 'rebased'
status check for the branches you are trying to push to.

You need to run the script at least once with an unpushed commits
in order for the 'rebased' status check to appear as an option
to enable in your github projects configuration.

Reviewers: jyknight

Subscribers: llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D67772